python_requirements.txt
python-docx>=1.2.0
langchain-community==0.0.33
langchain-openai==0.1.7
openai==1.30.1
unstructured==0.13.3
unstructured-inference==0.8.3
PyMuPDF==1.24.2
python-pptx==0.6.23
sentence-transformers==2.7.0
faiss-cpu==1.7.4
numpy==1.26.4
